version with moved Main Class
