+++ src/main/java/cz/inventi/kpj/spring/springintroduction/GreetingConfiguration.java
@@ -4,13 +4,12 @@ import cz.inventi.kpj.spring.springintroduction.impl.GreetingConstantRepositoryI
 import cz.inventi.kpj.spring.springintroduction.persistence.GreetingRepository;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
+
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
 
 @Configuration
 @EnableConfigurationProperties(GreetingConfigurationProperties.class)
-@ComponentScan(basePackages = "cz.inventi.kpj.spring.springintroduction")
